backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mail-parser==3.15.0
dkimpy==1.1.0
pyspf==2.0.14
checkdmarc==5.3.1
dnspython==2.6.1
beautifulsoup4==4.12.3
python-whois==0.9.4
httpx==0.27.0
upstash-redis==1.1.0
pydantic==2.6.0
python-multipart==0.0.9
